voice_bridge/requirements.txt
fastapi
uvicorn
websockets
deepgram-sdk
elevenlabs
python-dotenv
aiofiles